requirements.txt
langchain-ollama>=0.2.0
rich>=13.0.0
PyYAML>=6.0.0
